# Conflicts:
#	lib/src/pages/app_tabbar/app_tabbar_page.dart

+++ lib/src/pages/node/map3page/map3_node_create_contract_page.dart
@@ -377,8 +377,7 @@ Widget getHoldInNum(
     Function onPressFunction,
     {Function joinEnougnFunction,
     bool isMyself = false}) {
-  // todo: test_jison_0411
-  //List<int> suggestList = [];
+
   List<int> suggestList =
       contractNodeItem.contract.suggestQuantity.split(",").map((suggest) => int.parse(suggest)).toList();
 
@@ -411,7 +410,7 @@ Widget getHoldInNum(
           child: Row(
             children: <Widget>[
               Text(S.of(context).mortgage_hyn_num, style: TextStyle(fontWeight: FontWeight.w600, fontSize: 15)),
-              Text(S.of(context).mortgage_wallet_balance(walletName, FormatUtil.formatNumDecimal(balance)),
+              Text(S.of(context).mortgage_wallet_balance(walletName, FormatUtil.formatNumDecimal(balance.toInt().toDouble())),
                   style: TextStyle(color: Colors.grey[600])),
             ],
           ),
